llmagent/requirements.txt
llama-cpp-python
yfinance
python-dotenv
streamlit
textblob
pandas
google-generativeai
plotly
requests

